refactor: 참여자 관련 로직 리팩토링
