+++ src/main/java/com/backend/blooming/goal/domain/GoalTeam.java
@@ -39,9 +39,9 @@ public class GoalTeam extends BaseTimeEntity {
     @Column(nullable = false)
     private boolean deleted = false;
 
-    public GoalTeam(final User user, final Goal goal) {
+    public GoalTeam(final User user, final Goal goal, final Teams teams) {
         this.user = user;
         this.goal = goal;
-        this.goal.getTeams().add(this);
+        teams.getTeams().add(this);
     }
 }
